--- src/data/catcherList.json
@@ -1,99 +1,0 @@
-{
-  "data": {
-    "list": [
-      {
-        "backnum": "55",
-        "energybar": 55,
-        "energybarName": "55 점",
-        "gyear": "2024",
-        "hasFanpage": "N",
-        "hittype": "우투우타",
-        "mobilePlayerImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/50066_2024-03-06_130333.jpg",
-        "mobilePlayerImg1": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/50066_2024-03-06_130346.jpg",
-        "mobilePlayerImg2": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/50066_2024-03-06_130359.jpg",
-        "pcode": "50066",
-        "playerName": "강현우",
-        "playerPrvwImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/50066_2024-03-06_130258.jpg",
-        "position": "포수",
-        "rank": 38,
-        "rankName": "38 위",
-        "teamName": "KT"
-      },
-      {
-        "backnum": "97",
-        "energybar": 10,
-        "energybarName": "10 점",
-        "gyear": "2024",
-        "hasFanpage": "N",
-        "hittype": "우투우타",
-        "mobilePlayerImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/54097_2024-03-06_130453.jpg",
-        "mobilePlayerImg1": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/54097_2024-03-06_130527.jpg",
-        "mobilePlayerImg2": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/54097_2024-03-06_130543.jpg",
-        "pcode": "54097",
-        "playerName": "김민석",
-        "playerPrvwImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/54097_2024-03-06_130416.jpg",
-        "position": "포수",
-        "rank": 59,
-        "rankName": "59 위",
-        "teamName": "KT"
-      },
-      {
-        "backnum": "44",
-        "energybar": 19,
-        "energybarName": "19 점",
-        "gyear": "2024",
-        "hasFanpage": "N",
-        "hittype": "우투좌타",
-        "mobilePlayerImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/62558_2024-03-06_130743.jpg",
-        "mobilePlayerImg1": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/62558_2024-03-06_130858.jpg",
-        "mobilePlayerImg2": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/62558_2024-03-06_130810.jpg",
-        "pcode": "62558",
-        "playerName": "김준태",
-        "playerPrvwImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/62558_2024-03-06_130605.jpg",
-        "position": "포수",
-        "rank": 47,
-        "rankName": "47 위",
-        "teamName": "KT"
-      },
-      {
-        "backnum": "22",
-        "energybar": 112,
-        "energybarName": "112 점",
-        "gyear": "2024",
-        "hasFanpage": "Y",
-        "hittype": "우투우타",
-        "mobilePlayerImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/78548_2024-03-06_131013.jpg",
-        "mobilePlayerImg1": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/78548_2024-03-06_131105.jpg",
-        "mobilePlayerImg2": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/78548_2024-03-06_131129.jpg",
-        "pcode": "78548",
-        "playerName": "장성우",
-        "playerPrvwImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/78548_2024-03-06_130913.jpg",
-        "position": "포수",
-        "rank": 28,
-        "rankName": "28 위",
-        "teamName": "KT"
-      },
-      {
-        "backnum": "42",
-        "energybar": 453,
-        "energybarName": "453 점",
-        "gyear": "2024",
-        "hasFanpage": "N",
-        "hittype": "우투우타",
-        "mobilePlayerImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/68089_2024-03-08_104716.jpg",
-        "mobilePlayerImg1": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/68089_2024-03-06_131351.jpg",
-        "mobilePlayerImg2": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/68089_2024-03-06_131402.jpg",
-        "pcode": "68089",
-        "playerName": "조대현",
-        "playerPrvwImg": "https://wizzap.ktwiz.co.kr/files/playerImg/ktImg/68089_2024-03-08_104729.jpg",
-        "position": "포수",
-        "rank": 12,
-        "rankName": "12 위",
-        "teamName": "KT"
-      }
-    ]
-  },
-  "sVer": "2.0.1",
-  "status": 0,
-  "sName": "wizzap-server-api-v2"
-}
