--- src/data/gallery.json
@@ -1,226 +1,0 @@
-{
-  "gallery": {
-    "list": [
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190205,
-        "artcSubTitle": "이강철감독!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913140431.ed0-649a913933f9.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726200000000,
-        "regr": "helixtech1",
-        "startDttm": 1726200000000,
-        "totalPage": 4055,
-        "updDttm": 1726203840000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 2
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190204,
-        "artcSubTitle": "심우준선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913140431.ea5-ba37bad9d304.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726200000000,
-        "regr": "helixtech1",
-        "startDttm": 1726200000000,
-        "totalPage": 4055,
-        "updDttm": 1726203840000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 0
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190203,
-        "artcSubTitle": "김상수선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913140431.c93-2e99867d2bcd.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726200000000,
-        "regr": "helixtech1",
-        "startDttm": 1726200000000,
-        "totalPage": 4055,
-        "updDttm": 1726203840000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 0
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190201,
-        "artcSubTitle": "배정대선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913101243.491-f0e347375746.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726189200000,
-        "regr": "helixtech1",
-        "startDttm": 1726189200000,
-        "totalPage": 4055,
-        "updDttm": 1726189920000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 3
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190197,
-        "artcSubTitle": "배정대선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913101127.19e-eeb19505c600.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726188900000,
-        "regr": "helixtech1",
-        "startDttm": 1726188900000,
-        "totalPage": 4055,
-        "updDttm": 1726189860000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 2
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190195,
-        "artcSubTitle": "배정대선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913101127.9fe-1e277a03b8bb.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726188900000,
-        "regr": "helixtech1",
-        "startDttm": 1726188900000,
-        "totalPage": 4055,
-        "updDttm": 1726189860000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 2
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190193,
-        "artcSubTitle": "쿠에바스선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913093235.1d3-3e190d4ede6f.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726187700000,
-        "regr": "helixtech1",
-        "startDttm": 1726187700000,
-        "totalPage": 4055,
-        "updDttm": 1726187520000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 3
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190192,
-        "artcSubTitle": "소형준선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913093235.60e-378b908493ec.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726187700000,
-        "regr": "helixtech1",
-        "startDttm": 1726187700000,
-        "totalPage": 4055,
-        "updDttm": 1726187520000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 7
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190191,
-        "artcSubTitle": "소형준선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913093235.fe4-020d5878a02c.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726187700000,
-        "regr": "helixtech1",
-        "startDttm": 1726187700000,
-        "totalPage": 4055,
-        "updDttm": 1726187520000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 6
-      },
-      {
-        "artcNextSeq": 0,
-        "artcPrevSeq": 0,
-        "artcSeq": 190190,
-        "artcSubTitle": "로하스선수!",
-        "artcTitle": "9월 12일 NC와의 경기",
-        "boardCatSeq": 1,
-        "boardCode": "006",
-        "contentsDate": "2024-09-12",
-        "delYn": "N",
-        "endDttm": 253402225200000,
-        "imgFilePath": "https://wizzap.ktwiz.co.kr/files/article/2024/09/13/20240913093235.488-431cfa8da068.jpg",
-        "maxArticlePerPage": 0,
-        "regDttm": 1726187700000,
-        "regr": "helixtech1",
-        "startDttm": 1726187700000,
-        "totalPage": 4055,
-        "updDttm": 1726187520000,
-        "updr": "helixtech1",
-        "useYn": "Y",
-        "viewCnt": 1
-      }
-    ]
-  }
-}
