+++ src/data/game/boxScoreHeaders.ts
