--- src/data/recentMatches.json
@@ -1,92 +1,0 @@
-[
-  {
-    "displayDate": "20240911",
-    "game": "prev",
-    "gameDate": 20240911,
-    "gday": 11,
-    "gmkey": "20240911NCKT0",
-    "gmonth": 9,
-    "gtime": "18:30",
-    "gyear": "2024",
-    "home": "KT",
-    "homeDecision": "W",
-    "homeDecisionPitcher": "엄상백",
-    "homeFullname": "KT 위즈",
-    "homeKey": "KT",
-    "homeLogo": "http://3.35.51.214/api/static/KT.png",
-    "homeScore": 2,
-    "homeStarter": "엄상백",
-    "matchTeamCode": "NC",
-    "matchTeamName": "NC",
-    "outcome": "승",
-    "stadium": "수원",
-    "stadiumKey": "SW",
-    "status": "3",
-    "visit": "NC",
-    "visitDecision": "L",
-    "visitDecisionPitcher": "이재학",
-    "visitFullname": "NC 다이노스",
-    "visitKey": "NC",
-    "visitLogo": "http://3.35.51.214/api/static/NC.png",
-    "visitScore": 1,
-    "visitStarter": "이재학"
-  },
-  {
-    "displayDate": "20240912",
-    "game": "current",
-    "gameDate": 20240912,
-    "gday": 12,
-    "gmkey": "20240912NCKT0",
-    "gmonth": 9,
-    "gtime": "18:30",
-    "gyear": "2024",
-    "home": "KT",
-    "homeDecision": "W",
-    "homeDecisionPitcher": "쿠에바스",
-    "homeFullname": "KT 위즈",
-    "homeKey": "KT",
-    "homeLogo": "http://3.35.51.214/api/static/KT.png",
-    "homeScore": 10,
-    "homeStarter": "쿠에바스",
-    "matchTeamCode": "NC",
-    "matchTeamName": "NC",
-    "outcome": "승",
-    "stadium": "수원",
-    "stadiumKey": "SW",
-    "status": "3",
-    "visit": "NC",
-    "visitDecision": "L",
-    "visitDecisionPitcher": "임상현",
-    "visitFullname": "NC 다이노스",
-    "visitKey": "NC",
-    "visitLogo": "http://3.35.51.214/api/static/NC.png",
-    "visitScore": 4,
-    "visitStarter": "임상현"
-  },
-  {
-    "displayDate": "20240914",
-    "game": "next",
-    "gameDate": 20240914,
-    "gday": 14,
-    "gmkey": "20240914KTOB0",
-    "gmonth": 9,
-    "gtime": "17:00",
-    "gyear": "2024",
-    "home": "두산",
-    "homeFullname": "두산 베어스",
-    "homeKey": "OB",
-    "homeLogo": "http://3.35.51.214/api/static/OB.png",
-    "homeStarter": "곽빈",
-    "matchTeamCode": "OB",
-    "matchTeamName": "두산",
-    "outcome": "",
-    "stadium": "잠실",
-    "stadiumKey": "JS",
-    "status": "1",
-    "visit": "KT",
-    "visitFullname": "KT 위즈",
-    "visitKey": "KT",
-    "visitLogo": "http://3.35.51.214/api/static/KT.png",
-    "visitStarter": "고영표"
-  }
-]
