--- src/data/teamRanking.json
@@ -1,12 +1,0 @@
-[
-  {
-    "game": 132,
-    "gyear": "2024",
-    "rank": 4,
-    "rankName": "4위",
-    "teamCode": "KT",
-    "teamName": "KT",
-    "wldName": "65승 65패 2무",
-    "wra": "0.500"
-  }
-]
