+++ src/pages/ticket/Entrance.tsx
@@ -1,9 +1,9 @@
-import { StyledImage } from "@styles/ImageStyle.style";
+import { FullContentImage } from "@styles/common.style";
 
 const Entrance = () => {
   return (
     <>
-      <StyledImage
+      <FullContentImage
         src="https://wizzap.ktwiz.co.kr/files/resource/2024/03/19/20240319090539.d09-c6139b4f2910.png"
         alt="입장 시간 및 방법"
       />
