+++ src/pages/ticket/Store.tsx
@@ -1,9 +1,9 @@
-import { StyledImage } from "@styles/ImageStyle.style";
+import { FullContentImage } from "@styles/common.style";
 
 const Store = () => {
   return (
     <>
-      <StyledImage
+      <FullContentImage
         src="https://wizzap.ktwiz.co.kr/files/resource/2023/03/28/20230328173650.914-513d0b606ab1.png"
         alt="층별 편의시설 안내"
       />
