--- src/styles/ImageStyle.style.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const StyledImage = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-`;
