--- src/styles/Ranking.style.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const StyledArticle = styled.article`
-  width: 100%;
-  height: 435px;
-  margin-top: 45px;
-`;
