--- src/styles/WizParkGuide.style.ts
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-export const TitleImg = styled.img`
-  width: 342px;
-  height: 61px;
-`;
-
-export const TitleDiv = styled.div`
-  text-align: center;
-  margin-bottom: 30px;
-`;
-
-export const DescImg = styled.img`
-  width: 100%;
-`;
