--- src/styles/about.style.ts
@@ -1,100 +1,0 @@
-// .section1-container {
-//   margin-bottom: 100px;
-// }
-
-// .section2-container {
-//   height: 482px;
-// }
-
-// .img2 {
-//   float: right;
-//   height: auto;
-//   padding-left: 25px;
-//   text-align: center;
-// }
-
-// .section3-container {
-//   height: 300px;
-//   margin-bottom: 100px;
-// }
-
-// .img3 {
-//   float: left;
-//   height: auto;
-//   padding-right: 25px;
-//   text-align: center;
-// }
-
-// .title {
-//   font-size: 145%;
-//   color: #f03a3c;
-//   font-weight: 600;
-//   text-align: match-parent;
-// }
-
-// .title:after {
-//   content: "";
-//   display: block flex;
-//   width: 45px; /* 밑줄 길이 조절 */
-//   border-bottom: 3px solid #f03a3c;
-//   margin-top: 30px;
-// }
-
-// .text {
-//   font-size: 85%;
-//   margin-bottom: 80px;
-// }
-
-// .underline {
-//   color: #f03a3c;
-// }
-
-// .banner {
-//   margin-bottom: 70px;
-//   width: 100%;
-// }
-
-// /* History 컴포넌트 */
-// .history-section {
-//   white-space: pre-wrap;
-//   width: 430px;
-//   color: #111111;
-//   position: relative;
-// }
-
-// .ballIcon {
-//   width: 45px;
-//   height: 45px;
-//   position: absolute;
-//   top: 25px;
-//   left: -18px;
-// }
-
-// .year {
-//   font-size: 20px;
-//   height: 20px;
-// }
-
-// .content {
-//   /* position: relative; */
-//   background-color: #f9f9f9;
-//   border: 1px solid #ddd;
-//   font-size: 20px;
-//   color: #5b5a5a;
-//   padding: 38px 28px;
-
-//   /* 위에 빨간줄 */
-//   border-top: 4px solid red;
-//   background-color: #f5f5f5;
-// }
-
-// .history-container {
-//   float: right;
-// }
-
-/* .divider {
-  position: relative;
-  width: 3px; 
-  height: 3651px; 
-  background-color: black; 
-} */
